+++ bevy_asset_loader/tests/ui_with_render_feature/key_and_any_other_attribute.rs
@@ -28,3 +28,10 @@ struct AssetCollectionAndKey {
     #[asset(key = "test")]
     test: Handle<TextureAtlas>,
 }
+
+// This combination is allowed for optional dynamic assets
+#[derive(AssetCollection)]
+struct OptionalDynamic {
+    #[asset(key = "test", dsad)]
+    test: Handle<TextureAtlas>,
+}